sparseCV_L0: cross-validation functions. For internal package use only.

sparseCV_L0: cross-validation functions. For internal package use only.

sparseL0Tn_iht(
  data,
  tune.grid,
  hoso = "hoso",
  nfolds = "K",
  juliaFnPath = "/Users/gabeloewinger/Desktop/Research Final/Sparse Multi-Study/",
  trainingStudy = NA,
  messageInd = FALSE,
  LSitr = 50,
  LSspc = 1,
  maxIter = 2500
)

Arguments

data

Matrix with outcome and design matrix

tune.grid

A data.frame of tuning values

hoso

String specifying tuning type

nfolds

String or integer specifying number of folds

juliaFnPath

String specifying path to Julia binary

trainingStudy

Integer specifying index of training study

messageInd

Boolean for message printing

LSitr

Integer specifying do <LSitr> local search iterations on parameter values where we do actually do LS; NA does no local search

LSspc

Integer specifying number of hyperparameters to conduct local search: conduct local search every <LSspc>^th iteration. NA does no local search

maxIter

Integer specifying max iterations of coordinate descent